locations/serializers/provider_serializer.py
from rest_framework import serializers
from locations.models import Provider


class ProviderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Provider
        fields = "__all__"
