requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
dj-database-url==0.5.0
Django==4.0.2
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
